@@ -15,36 +15,36 @@
   },
   "devDependencies": {
     "@astrojs/sitemap": "^3.2.1",
-    "@astrojs/tailwind": "^6.0.0",
-    "@eslint/js": "^9.9.1",
-    "@fontsource-variable/jetbrains-mono": "^5.0.22",
-    "@iconify-json/fa-brands": "^1.1.9",
-    "@iconify-json/icon-park-outline": "^1.1.16",
-    "@iconify-json/mdi": "^1.1.68",
-    "@iconify-json/ph": "^1.1.14",
+    "@astrojs/tailwind": "^5.1.3",
+    "@eslint/js": "^9.21.0",
+    "@fontsource-variable/jetbrains-mono": "^5.1.2",
+    "@iconify-json/fa-brands": "^1.2.1",
+    "@iconify-json/icon-park-outline": "^1.2.2",
+    "@iconify-json/mdi": "^1.2.3",
+    "@iconify-json/ph": "^1.2.2",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/typography": "^0.5.15",
+    "@tailwindcss/typography": "^0.5.16",
     "@types/eslint": "^9.6.1",
-    "@typescript-eslint/eslint-plugin": "^8.3.0",
-    "@typescript-eslint/parser": "^8.3.0",
-    "astro-expressive-code": "^0.36.0",
-    "astro-icon": "^1.1.1",
+    "@typescript-eslint/eslint-plugin": "^8.24.1",
+    "@typescript-eslint/parser": "^8.24.1",
+    "astro-expressive-code": "^0.39.0",
+    "astro-icon": "^1.1.5",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.9.1",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-astro": "^1.2.3",
-    "eslint-plugin-prettier": "^5.2.1",
-    "globals": "^15.9.0",
-    "hast-util-select": "^6.0.2",
-    "postcss": "^8.4.41",
-    "prettier": "^3.3.3",
+    "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-astro": "^1.3.1",
+    "eslint-plugin-prettier": "^5.2.3",
+    "globals": "^16.0.0",
+    "hast-util-select": "^6.0.4",
+    "postcss": "^8.5.3",
+    "prettier": "^3.5.2",
     "prettier-plugin-astro": "^0.14.1",
-    "prettier-plugin-tailwindcss": "^0.6.6",
+    "prettier-plugin-tailwindcss": "^0.6.11",
     "rehype-autolink-headings": "^7.1.0",
     "rehype-external-links": "^3.0.0",
     "rehype-slug": "^6.0.0",
-    "sharp": "0.33.2",
-    "tailwindcss": "^3.4.10",
-    "typescript-eslint": "^8.3.0"
+    "sharp": "0.33.5",
+    "tailwindcss": "^3.4.17",
+    "typescript-eslint": "^8.24.1"
   }
 }
