@@ -23,7 +23,10 @@ export default defineConfig({
         if (styleVariants.length >= 2) {
           const baseTheme = styleVariants[0]?.theme;
           const altTheme = styleVariants.find((v) => v.theme.type !== baseTheme?.type)?.theme;
-          if (theme === baseTheme || theme === altTheme) return `[data-theme='${theme.type}']`;
+
+          if (theme === baseTheme || theme === altTheme) {
+            return `[data-theme='${theme.type}']`;
+          }
         }
         // return default selector
         return `[data-theme="${theme.name}"]`;
